+++ BookDTO.cs
