+++ BookRepository.cs
@@ -1,46 +1,102 @@
 ﻿using System;
-using CrudConsole;
+using System.Collections.Generic;
+using System.Data.SqlClient;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Dapper;
+using Microsoft.EntityFrameworkCore;
+
 // the x service knows how to handle x
-public class CRUDLogic
+// when setting up database naming schemes, you can say all camelcase is converted to SQL conventions
+
+namespace CrudConsole
 {
-    // These 4 methods become part of the database class
-
-    private string filePath = FilePaths.filePath;
-
-    public void DisplayBooks()
+    // repository pattern
+    public interface IRepository
     {
-        Console.WriteLine("Current books in library:");
-        Interface.PrintLines(FileInteracter.ReadLinesFromFile(filePath), filePath);
+        // create a test for reading database
+        // public unneccessary 
+        public void ReadDatabase();
+        public void AddBook();
+        public void EditBook();
+        public void RemoveBook();
     }
 
-    public void AddBooks()
+    public class LibraryContext : DbContext
     {
-        List<String> books = FileInteracter.ReadLinesFromFile(filePath).ToList();
-        List<String> newBooks = BookOperations.ConvertBookListToJSON(BookOperations.GetUsersListOfBooks());
+        public DbSet<Book> simple_library { get; set; }
 
-        books.AddRange(newBooks);
+        public static String connectionString = "Server=localhost;User ID=sa;Password=9n8kZ81J0iuB;Initial Catalog=SIMPLE_LIBRARY;Integrated Security=false;TrustServerCertificate=True";
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+            optionsBuilder.UseSqlServer(connectionString);
+        }
 
-        FileInteracter.WriteLinesToFile(books, filePath);
     }
 
-    public void RemoveBook()
+    public class DapperDBConnect : IDatabase
     {
-        int chosenBook = BookOperations.GetIndexOfBookToModify("remove");
-        List<string> lines = FileInteracter.ReadLinesFromFile(filePath).ToList();
-        lines.RemoveAt(chosenBook);
-        FileInteracter.WriteLinesToFile(lines, filePath);
-        Console.WriteLine("Book removed.");
-        DisplayBooks();
+        public static String connectionString = "Server=localhost;User ID=sa;Password=9n8kZ81J0iuB;Initial Catalog=SIMPLE_LIBRARY;Integrated Security=false;TrustServerCertificate=True";
+
+        public void ReadDatabase()
+        {
+            using (var connection = new SqlConnection(connectionString))
+            {
+                var sql = "SELECT * FROM dbo.simple_library";
+
+                var books = connection.Query<Book>(sql).ToList();
+
+                foreach (Book book in books)
+                {
+                    Console.WriteLine(book.ToString());
+                }
+            }
+        }
+
+        public void AddBook()
+        {
+            using (var connection = new SqlConnection(connectionString))
+            {
+                // get value from user
+
+                var sql = "INSERT INTO dbo.simple_library";
+
+                var books = connection.Query<Book>(sql).ToList();
+
+                foreach (Book book in books)
+                {
+                    Console.WriteLine(book.ToString());
+                }
+            }
+        }
+
+        public void EditBook()
+        {
+
+        }
+
+        public void RemoveBook()
+        {
+
+        }
+
     }
 
-    public void UpdateBook()
+    public class EntityDBConnect
     {
-        int bookIndex = BookOperations.GetIndexOfBookToModify("update");
-        List<string> lines = FileInteracter.ReadLinesFromFile(filePath).ToList();
-        string book = lines[bookIndex];
-        string updatedBook = BookOperations.ChangeBookProperties(book);
-        lines[bookIndex] = updatedBook;
-        FileInteracter.WriteLinesToFile(lines, filePath);
 
+        public static void QueryBooks()
+        {
+            using (var context = new LibraryContext())
+            {
+                Console.WriteLine("Testing...");
+                var books = context.simple_library.ToList();
+                foreach (var book in books)
+                {
+                    Console.WriteLine(book);
+                }
+            }
+        }
     }
 }
